Move all Latex helpers to one file
