--- shared/latex_div.py
@@ -1,6 +1,0 @@
-from bokeh.models.widgets import Div
-
-class LatexDiv(Div):
-    __javascript__ = ["https://cdnjs.cloudflare.com/ajax/libs/KaTeX/0.6.0/katex.min.js"]
-    __css__ = ["https://cdnjs.cloudflare.com/ajax/libs/KaTeX/0.6.0/katex.min.css"]
-    __implementation__ = "latex_div.ts"
